+++ src/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Changed
 
+- Groups rows can now be deleted through ui.
 - Added version to history to make tracking updates faster.
 - Fully replaced dndkit with newer version. Groups now supports multiple rows.
 - Replacing dndkit legacy with new version.